feat: from_of
