+++ examples/index-juego.js
@@ -1,4 +1,5 @@
-import { fromEvent } from "rxjs";
+import { fromEvent, Subject } from "rxjs";
+import WORDS_LIST from './wordsList.json'
 
 /**
  * Captura los elementos de esa clase en el html
@@ -10,9 +11,22 @@ const letterRow = document.getElementsByClassName('letter-row');
  */
 const onKeyDown$ = fromEvent(document, "keydown")
 
+/**
+ * Observable subject
+ */
+const userWinOrLoose$ = new Subject();
+
 /**Posición en x */
 let letterX = 0;
 let letterY = 0;
+let userAnswer = [];
+
+/**Obtener una palabra de la lista aleatoriamente */
+const getRandomWord = () => WORDS_LIST[Math.floor(Math.random() * WORDS_LIST.length)];
+
+let rigthtWord = getRandomWord();
+console.log(rigthtWord)
+
 
 /**
  * Observador que permite insertar letras
@@ -23,21 +37,36 @@ const insertLetter = {
         if (preesedKey.length === 1 && preesedKey.match(/[a-z]/i)) {
             let letterBox = Array.from(letterRow)[letterX].children[letterY];
             letterBox.textContent = preesedKey;
+            userAnswer.push(preesedKey)
             letterY++;
         }
     }
 }
 
+/**
+ * Observador encargado de borrar la letra ingresada 
+ */
 const deleteLetter = {
     next: (event) => {
         const preesedKey = event.key;
-        if(preesedKey === 'Backspace') {
+        if (preesedKey === 'Backspace') {
             letterY--;
             let letterBox = Array.from(letterRow)[letterX].children[letterY];
             letterBox.textContent = "";
+            userAnswer.pop()
 
         }
-        console.log(preesedKey)
+    }
+}
+
+/**verifica la palabra que ingresa */
+const checkWord = {
+    next: (event) => {
+        if (event.key === 'Enter') {
+            if (userAnswer.join("") === rigthtWord) {
+                userWinOrLoose$.next()
+            }
+        }
     }
 }
 
@@ -47,6 +76,17 @@ const deleteLetter = {
 onKeyDown$.subscribe(insertLetter);
 
 /**
- * Subscrmirme al otro observable
+ * Subcribirme al otro observable
  */
 onKeyDown$.subscribe(deleteLetter);
+
+/**verificar que sea igual la palabra */
+onKeyDown$.subscribe(checkWord);
+
+userWinOrLoose$.subscribe(() => {
+    
+    let lettersRowsWinned = Array.from(letterRow)[letterX]
+    for (let i = 0; i < 5; i++) {
+        lettersRowsWinned.children[i].classList.add('letter-green');
+    }
+})
